requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.0
django-environ==0.9.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2
sqlparse==0.4.2
whitenoise==6.2.0
